backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
psycopg2==2.8.6
requests==2.24.0